[AISW-154328] [QNN-EP] Add MatMulNBits translation for GPU
  Add support for translation of MatMulNBits contrib op to
  QNN with FullyConnected operation with INT4 BlockQuantized weights

Implementation details:
 - Translate MatMulNBits to FullyConnected in OpBuilder
 - Support QNN_QUANTIZATION_ENCODING_BLOCK for INT4 weights
 - Pass INT4 weights and quant params as BlockQuantization encoding params in QNN

Testing:
 - Added new unit tests for MNB -> QNN-GPU
 - Validated all OnnxRuntime tests
 - Validated the following LLMs through Olive and ORT-GenAI execution flow
   - LlaMA3.2 1B
   - Qwen2.5
   - DeepSeek-R1-Qwen 1.5b
   - Phi3.5-mini-instruct

Motivation and Context:

 LLMs with INT4 quantization pass in Olive will generate a model with MatMulMBits contrib ops.
 To run these ops via QNN-EP, MatMulNBits is translated to QNN FullyConnected op with INT4 weights.

+++ onnxruntime/core/providers/qnn/builder/qnn_quant_params_wrapper.cc
@@ -192,7 +192,7 @@ QnnQuantParamsWrapper::QnnQuantParamsWrapper(
     block_encoding_scale_offsets_data_ = std::make_unique<char[]>(be_scale_offsets_num_bytes + be_scale_offsets_align);
     Qnn_ScaleOffset_t* block_encoding_scale_offsets_data_aligned = ALIGN_PTR_UP(block_encoding_scale_offsets_data_.get(), be_scale_offsets_align, Qnn_ScaleOffset_t*);
 
-    for (size_t i = 0; i < static_cast<uint32_t>(num_blocks_); i++) {
+    for (uint32_t i = 0; i < num_blocks_; i++) {
       block_encoding_scale_offsets_data_aligned[i].offset = offsets[i];
       block_encoding_scale_offsets_data_aligned[i].scale = scales[i];
     }
